requirements.txt
celery==5.1.2
Django==3.2.8
django-celery-beat==2.2.1
django-crispy-forms==1.13.0
psycopg2-binary==2.9.1
gunicorn==20.1.0
redis==3.5.3
django-extensions
